+++ rtic-monotonics/CHANGELOG.md
@@ -7,7 +7,7 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ## Unreleased
 
-- Added Silabs (EFR32, EFM32) support using the inbuilt RTCC / LETIMER peripherals
+## v2.2.0 - 2026-07-08
 
 ### Changed
 
@@ -18,6 +18,7 @@ For each category, *Added*, *Changed*, *Fixed* add new entries at the top!
 
 ### Added
 
+- Silabs (EFR32, EFM32) support using the inbuilt RTCC / LETIMER peripherals
 - nRF9151 and nRF9161 support (RTC and TIMER monotonics), backed by `nrf9120-pac`
 - Secure and non-secure variants for all nRF91 chips: `nrf9160-s`/`nrf9160-ns`, `nrf9151-s`/`nrf9151-ns`, `nrf9161-s`/`nrf9161-ns`
 - A lot of STM32 chip feature flags